requirements.txt
Flask==2.3.3
google-generativeai==0.5.2
python-dotenv==1.0.1
pyttsx3==2.90; sys_platform != 'linux'
